+++ xdump/cli/dump.py
@@ -3,10 +3,12 @@ import zipfile
 
 import click
 
+from .utils import apply_decorators, import_backend
 
-@click.group()
+
+@click.group(name='xdump')
 @click.version_option()
-def xdump():
+def dump():
     pass
 
 
@@ -34,7 +36,7 @@ if sys.version_info[0] == 3:
 
 
 DEFAULT_PARAMETERS = [
-    xdump.command(),
+    dump.command(),
     click.option('-D', '--dbname', required=True, help='database to dump'),
     click.option('-o', '--output', required=True, help='output file name'),
     click.option('-f', '--full', help='table name to be fully dumped. Could be used multiple times', multiple=True),
@@ -52,31 +54,16 @@ DEFAULT_PARAMETERS = [
     ),
     click.option('--dump-schema', help='include / exclude the schema from the dump', default=True, type=bool),
     click.option('--dump-data', help='include / exclude the data from the dump', default=True, type=bool),
+    click.option('-v', '--verbosity', help='verbosity level', default=0, count=True, type=click.IntRange(0, 2)),
 ]
-
-
-def apply_decorators(decorators):
-
-    def wrapper(func):
-        for decorator in reversed(decorators):
-            func = decorator(func)
-        return func
-
-    return wrapper
 
 
 def command(func):
     return apply_decorators(DEFAULT_PARAMETERS)(func)
 
 
-def import_backend(path):
-    module_name, class_name = path.rsplit('.', 1)
-    module = __import__(module_name, fromlist=[class_name])
-    return getattr(module, class_name)
-
-
 def base_dump(backend_path, user, password, host, port, dbname, output, full, partial, compression, dump_schema,
-              dump_data):
+              dump_data, verbosity):
     compression = COMPRESSION_MAPPING[compression]
 
     click.echo('Dumping ...')
@@ -84,7 +71,7 @@ def base_dump(backend_path, user, password, host, port, dbname, output, full, pa
 
     backend_class = import_backend(backend_path)
 
-    backend = backend_class(dbname=dbname, host=host, port=port, user=user, password=password)
+    backend = backend_class(dbname=dbname, host=host, port=port, user=user, password=password, verbosity=verbosity)
     backend.dump(
         output, full_tables=full, partial_tables=partial, compression=compression, dump_schema=dump_schema,
         dump_data=dump_data
@@ -97,12 +84,12 @@ def base_dump(backend_path, user, password, host, port, dbname, output, full, pa
 @click.option('-W', '--password', help='password for the DB connection')
 @click.option('-H', '--host', default='127.0.0.1', help='database server host or socket directory')
 @click.option('-P', '--port', default='5432', help='database server port number')
-def postgres(user, password, host, port, dbname, output, full, partial, compression, dump_schema, dump_data):
+def postgres(user, password, host, port, dbname, output, full, partial, compression, dump_schema, dump_data, verbosity):
     base_dump('xdump.postgresql.PostgreSQLBackend', user, password, host, port, dbname, output, full, partial,
-              compression, dump_schema, dump_data)
+              compression, dump_schema, dump_data, verbosity)
 
 
 @command
-def sqlite(dbname, output, full, partial, compression, dump_schema, dump_data):
+def sqlite(dbname, output, full, partial, compression, dump_schema, dump_data, verbosity):
     base_dump('xdump.sqlite.SQLiteBackend', None, None, None, None, dbname, output, full, partial, compression,
-              dump_schema, dump_data)
+              dump_schema, dump_data, verbosity)
